feat(129): Cleanup from code review
