+++ packages/frontend/src/components/ui/buttons/ActionButton.vue
@@ -98,7 +98,7 @@ const props = defineProps({
 
 /* === SIZE VARIANTS === */
 .action-btn--small {
-    padding: 2px var(--space-sm);
+    padding: var(--btn-padding-sm);
     font-size: var(--font-size-12);
     min-height: var(--btn-min-height-sm);
 }
