+++ packages/frontend/src/constants/filterBar.js
@@ -1,0 +1,9 @@
+export const FILTER_BAR_TAG_PICKER_MODES = {
+    FLAT: 'flat',
+    CASCADE: 'cascade',
+}
+
+// Label used to identify the special-filters tag group within tagGroups.
+// Callers that produce tagGroups must use this constant for the group label
+// so FilterBar can locate and render it correctly.
+export const FILTER_SPECIAL_TAG_GROUP_LABEL = 'Special Tags'
